+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ### Fixed
 - Dependency conflicts preventing app from running on fresh setup
 - Ensured npm install && npm run dev works correctly
+- Made time range selector fully functional
 
 ### Added
 - Add latency, error rate, and request rate to service cards
 - Historical data based selected time range
 - Service dropdown added for alert table and it is persistence with URL state.
+- Prepend live stream alerts with highlight animation on alert table.